+++ Controllers/UsuarioController.cs
@@ -3,19 +3,21 @@ using CursosAPI.Services;
 using Models;
 namespace CursosAPI.Controllers
 {
-   [Route("api/[controller]")]
-   [ApiController]
-   public class UsuarioController : ControllerBase
-   {
-    private static List<Usuario> usuarios = new List<Usuario>();
+    [Route("api/[controller]")]
+    [ApiController]
+    public class UsuarioController : ControllerBase
+    {
 
-    private readonly IUsuarioService _service;
 
-    public UsuarioController(IUsuarioService service)
+        private readonly IUsuarioService _service;
+        private readonly IConfiguration _configuration;
+
+        public UsuarioController(IUsuarioService service, IConfiguration configuration)
         {
             _service = service;
+            _configuration = configuration;
         }
-    
+
         [HttpGet]
         public async Task<ActionResult<List<Usuario>>> GetUsuarios()
         {
@@ -41,9 +43,14 @@ namespace CursosAPI.Controllers
         }
 
 
-       [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateUsuario(int id, UpdateUsuarioDTO updatedUsuario)
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateUsuario(int id, UpdateUsuarioDTO updatedUsuario, [FromHeader(Name = "X-Admin-Key")] string? apiKey)
         {
+            var adminApiKey = _configuration["AdminSettings:ApiKey"];
+            if (string.IsNullOrEmpty(apiKey) || apiKey != adminApiKey)
+            {
+                return Unauthorized("Acceso denegado: ApiKey inválida.");
+            }
             var existingUsuario = await _service.GetByIdAsync(id);
             if (existingUsuario == null)
             {
@@ -59,18 +66,24 @@ namespace CursosAPI.Controllers
             await _service.UpdateAsync(existingUsuarioDTO, existingUsuario.Id);
             return NoContent();
         }
-  
-       [HttpDelete("{id}")]
-       public async Task<IActionResult> DeleteUsuario(int id)
-       {
-           var usuario = await _service.GetByIdAsync(id);
-           if (usuario == null)
-           {
-               return NotFound();
-           }
-           await _service.DeleteAsync(id);
-           return NoContent();
-       }
 
-   }
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteUsuario(int id, [FromHeader(Name = "X-Admin-Key")] string? apiKey)
+        {
+
+            var adminApiKey = _configuration["AdminSettings:ApiKey"];
+            if (string.IsNullOrEmpty(apiKey) || apiKey != adminApiKey)
+            {
+                return Unauthorized("Acceso denegado: ApiKey inválida.");
+            }
+            var usuario = await _service.GetByIdAsync(id);
+            if (usuario == null)
+            {
+                return NotFound();
+            }
+            await _service.DeleteAsync(id);
+            return NoContent();
+        }
+
+    }
 }
